refactor: QueryBuilder 리팩토링
* DataTypeMapper 생성하여 column type 매핑할 수 있도록 변경
* StringUtils 생성하여 범용성있게 사용하도록 변경
