+++ src/main/java/domain/Person.java
@@ -1,4 +1,4 @@
-package persistence.sql.ddl;
+package domain;
 
 import jakarta.persistence.*;
