+++ src/test/java/persistence/sql/ddl/QueryBuilderTest.java
@@ -1,5 +1,6 @@
 package persistence.sql.ddl;
 
+import domain.Person;
 import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
